src/data_loader/requirements.txt
requests==2.32.3
psycopg2-binary
numpy==1.24.4
pandas==2.0.3